refactor: Update imports to new component location
In this commit:
- Updated import paths for components to their new location after moving them from '@/components/'.

This change ensures that the project's import paths are consistent with the updated component location, making the codebase more maintainable and organized.
