+++ app/(main)/dogs/dog/[dogId]/page.tsx
@@ -1,7 +1,7 @@
 import getCurrentUser from "@/actions/getCurrentUser";
 import getDogsById from "@/actions/getDogsById";
 import getNearbyDogs from "@/actions/getNearbyDogs";
-import DogClient from "@/components/Dogs/DogClient";
+import DogClient from "@/app/(main)/dogs/_components/DogProfile";
 import { SIGN_IN_PATH } from "@/lib/constants";
 import { redirect } from "next/navigation";
 
@@ -15,22 +15,21 @@ const DogPage = async ({ params }: DogPageParams) => {
   const { dogId } = params;
   if (!dogId) throw new Error("No dog id found");
 
-  const currentUser = (await getCurrentUser()) || null;
+  const currentUser = await getCurrentUser();
   if (!currentUser) {
     return redirect(`${process.env.ORIGIN}${SIGN_IN_PATH}`);
   }
 
-  const [dog] =
-    (await getDogsById({
-      dogIdsToRetrieve: [dogId],
-      user: currentUser,
-    })) || null;
+  const [dog] = await getDogsById({
+    dogIdsToRetrieve: [dogId],
+    user: currentUser,
+  });
 
   if (!dog) {
     throw new Error("No dog found");
   }
 
-  const nearbyDogs = (await getNearbyDogs(dog.zip_code)) || null;
+  const nearbyDogs = await getNearbyDogs(dog.zip_code);
 
   return (
     <>
