+++ app/(main)/favorites/page.tsx
@@ -1,19 +1,19 @@
 import getCurrentUser from "@/actions/getCurrentUser";
 import EmptyState from "@/components/EmptyState";
 import getFavoriteDogs from "@/actions/getFavoriteDogs";
-import FavoriteDogs from "@/components/Dogs/FavoriteDogs";
+import FavoriteDogs from "@/app/(main)/favorites/_components/FavoriteDogs";
 import { SIGN_IN_PATH } from "@/lib/constants";
 import { redirect } from "next/navigation";
 
 const FavoriteDogsPage = async () => {
-  const currentUser = (await getCurrentUser()) || null;
+  const currentUser = await getCurrentUser();
 
   if (!currentUser) {
     <EmptyState title="You have been logged out. Redirecting to sign in" />;
     return redirect(`${process.env.ORIGIN}${SIGN_IN_PATH}`);
   }
 
-  const favoriteDogs = (await getFavoriteDogs()) || null;
+  const favoriteDogs = await getFavoriteDogs();
   if (!favoriteDogs) throw new Error("No favorite dogs found");
 
   if (favoriteDogs.length === 0) {
