+++ app/(main)/layout.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@/components/Navbar";
+import Navbar from "@/app/(main)/_components/Navbar/Navbar";
 import getCurrentUser from "@/actions/getCurrentUser";
 import { redirect } from "next/navigation";
 import { SIGN_IN_PATH } from "@/lib/constants";
@@ -10,9 +10,9 @@ interface SearchLayoutProps {
 const SearchLayout = async ({ children }: SearchLayoutProps) => {
   if (!children) throw new Error("No children found");
 
-  const currentUser = (await getCurrentUser()) || null;
+  const currentUser = await getCurrentUser();
   if (!currentUser) {
-    return redirect(SIGN_IN_PATH);
+    return redirect(`${process.env.ORIGIN}${SIGN_IN_PATH}`);
   }
 
   return (
