+++ app/page.tsx
@@ -3,7 +3,7 @@ import { SEARCH_PATH, SIGN_IN_PATH } from "@/lib/constants";
 import { redirect } from "next/navigation";
 
 const Home = async () => {
-  const currentUser = (await getCurrentUser()) || null;
+  const currentUser = await getCurrentUser();
   if (!currentUser) {
     return redirect(SIGN_IN_PATH);
   }
